Restructure service/ around one axis: subsystem folder vs single-type file
Make the module tree express a single rule — a folder holds a multi-concept
subsystem, a flat file holds exactly one type:

- Flatten `notification/` -> `notification.rs` and `retention/` ->
  `file_retention.rs`; both were folders wrapping one type behind a re-export
  `mod.rs` that earned nothing.
- Dissolve `security/` (a junk drawer of unrelated concerns): `session_keys`
  and `user_agent` become root files (peers of `avatar`/`infra`), and password
  handling — the one member with its own internal split — becomes `password/`
  (`service.rs` + `hasher.rs` + `strength.rs`).

Genuine subsystems keep their folders unchanged (`crypto/`, `engine/`,
`health/`, `detection/`, `sync/`, `webhook/`). The public facade in
`service/mod.rs` is the only external contract, so no call sites outside
`service/` change.

Full gate green (check / fmt / clippy / doc / test).
