--- crates/nvisy-server/src/service/security/mod.rs
@@ -1,14 +1,0 @@
-//! Security infrastructure services.
-//!
-//! This module provides authentication-related services including password
-//! handling, JWT secret key management, and user agent parsing.
-
-mod password;
-mod password_hasher;
-mod password_strength;
-mod session_keys;
-mod user_agent;
-
-pub use password::PasswordService;
-pub use session_keys::{SessionKeys, SessionKeysConfig};
-pub use user_agent::UserAgentParser;
